+++ package.json
@@ -57,6 +57,7 @@
     "@ngx-translate/core": "^12.1.2",
     "@ngx-translate/http-loader": "^4.0.0",
     "@stackblitz/sdk": "^1.3.0",
+    "codesandbox": "^2.1.14",
     "ajv": "^6.12.0",
     "extend": "^3.0.2",
     "file-saver": "^2.0.2",
