+++ src/app/core/code/files/delon-abc.module.ts
@@ -1,0 +1,66 @@
+export default `import { NgModule } from '@angular/core';
+
+// #region all modules
+import { AvatarListModule } from '@delon/abc/avatar-list';
+import { CountDownModule } from '@delon/abc/count-down';
+import { DatePickerModule } from '@delon/abc/date-picker';
+import { DownFileModule } from '@delon/abc/down-file';
+import { EllipsisModule } from '@delon/abc/ellipsis';
+import { ErrorCollectModule } from '@delon/abc/error-collect';
+import { ExceptionModule } from '@delon/abc/exception';
+import { FooterToolbarModule } from '@delon/abc/footer-toolbar';
+import { FullContentModule } from '@delon/abc/full-content';
+import { GlobalFooterModule } from '@delon/abc/global-footer';
+import { ImageModule } from '@delon/abc/image';
+import { LoadingModule } from '@delon/abc/loading';
+import { LodopModule } from '@delon/abc/lodop';
+import { NoticeIconModule } from '@delon/abc/notice-icon';
+import { NumberToChineseModule } from '@delon/abc/number-to-chinese';
+import { PageHeaderModule } from '@delon/abc/page-header';
+import { QRModule } from '@delon/abc/qr';
+import { QuickMenuModule } from '@delon/abc/quick-menu';
+import { ResultModule } from '@delon/abc/result';
+import { ReuseTabModule } from '@delon/abc/reuse-tab';
+import { SEModule } from '@delon/abc/se';
+import { SGModule } from '@delon/abc/sg';
+import { SidebarNavModule } from '@delon/abc/sidebar-nav';
+import { STModule } from '@delon/abc/st';
+import { SVModule } from '@delon/abc/sv';
+import { TagSelectModule } from '@delon/abc/tag-select';
+import { XlsxModule } from '@delon/abc/xlsx';
+import { ZipModule } from '@delon/abc/zip';
+
+const MODULES = [
+  ErrorCollectModule,
+  FooterToolbarModule,
+  SidebarNavModule,
+  DownFileModule,
+  ImageModule,
+  AvatarListModule,
+  EllipsisModule,
+  GlobalFooterModule,
+  ExceptionModule,
+  NoticeIconModule,
+  PageHeaderModule,
+  ResultModule,
+  TagSelectModule,
+  CountDownModule,
+  STModule,
+  ReuseTabModule,
+  FullContentModule,
+  XlsxModule,
+  ZipModule,
+  NumberToChineseModule,
+  LodopModule,
+  QuickMenuModule,
+  QRModule,
+  SVModule,
+  SEModule,
+  SGModule,
+  DatePickerModule,
+  LoadingModule,
+];
+
+@NgModule({ exports: MODULES })
+export class DemoDelonABCModule {}
+`;
