+++ src/app/core/code/files/delon-chart.module.ts
@@ -1,0 +1,43 @@
+export default `import { NgModule } from '@angular/core';
+
+// #region all modules
+
+import { G2BarModule } from '@delon/chart/bar';
+import { G2CardModule } from '@delon/chart/card';
+import { G2CustomModule } from '@delon/chart/custom';
+import { G2GaugeModule } from '@delon/chart/gauge';
+import { G2MiniAreaModule } from '@delon/chart/mini-area';
+import { G2MiniBarModule } from '@delon/chart/mini-bar';
+import { G2MiniProgressModule } from '@delon/chart/mini-progress';
+import { NumberInfoModule } from '@delon/chart/number-info';
+import { G2PieModule } from '@delon/chart/pie';
+import { G2RadarModule } from '@delon/chart/radar';
+import { G2SingleBarModule } from '@delon/chart/single-bar';
+import { G2TagCloudModule } from '@delon/chart/tag-cloud';
+import { G2TimelineModule } from '@delon/chart/timeline';
+import { TrendModule } from '@delon/chart/trend';
+import { G2WaterWaveModule } from '@delon/chart/water-wave';
+
+const MODULES = [
+  G2BarModule,
+  G2CardModule,
+  G2CustomModule,
+  G2GaugeModule,
+  G2MiniAreaModule,
+  G2MiniBarModule,
+  G2MiniProgressModule,
+  G2PieModule,
+  G2RadarModule,
+  G2TagCloudModule,
+  G2TimelineModule,
+  G2WaterWaveModule,
+  G2SingleBarModule,
+  NumberInfoModule,
+  TrendModule,
+];
+
+// #endregion
+
+@NgModule({ exports: MODULES })
+export class DemoDelonChartModule {}
+`;
