+++ src/app/core/code/files/polyfill.ts
@@ -1,0 +1,4 @@
+export default `
+import 'zone.js/dist/zone';
+import "core-js/proposals/reflect-metadata";
+`;
