+++ src/app/core/i18n/zh-CN.ts
@@ -62,6 +62,7 @@ export const ZHCN = {
   'app.component.config.title': '支持通用配置',
   'app.demo.copy': '复制代码',
   'app.demo.copied': '复制成功',
+  'app.demo.code-sandbox': '在 CodeSandbox 上打开',
   'app.demo.stackblitz': '在 StackBlitz 中打开',
   'sf.name': '名称',
   'sf.description': '这是一段描述',
